Created shared element registry
Moved shared element code out of NavigationMotion into registry
