+++ NavigationReactMobile/src/SharedElementContext.ts
@@ -1,3 +1,4 @@
+import SharedElementRegistry from './SharedElementRegistry';
 import * as React from 'react';
 
-export default React.createContext({ registerSharedElement: () => {}, unregisterSharedElement: () => {} });
+export default React.createContext(new SharedElementRegistry());
